+++ src/main/java/team03/mopl/domain/review/dto/ReviewResponse.java
@@ -1,11 +1,12 @@
 package team03.mopl.domain.review.dto;
 
+import java.math.BigDecimal;
 import team03.mopl.domain.review.entity.Review;
 
 public record ReviewResponse(
     String title,
     String comment,
-    float rating
+    BigDecimal rating
 ) {
 
   public static ReviewResponse from(Review review) {
