+++ src/main/java/team03/mopl/domain/review/dto/ReviewUpdateRequest.java
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.Digits;
 import jakarta.validation.constraints.NotBlank;
+import java.math.BigDecimal;
 
 public record ReviewUpdateRequest(
 
@@ -16,7 +17,7 @@ public record ReviewUpdateRequest(
     @DecimalMin(value = "0.0", message = "별점은 0.0 이상이어야 합니다")
     @DecimalMax(value = "5.0", message = "별점은 5.0 이하여야 합니다")
     @Digits(integer = 1, fraction = 1, message = "소수점 한자리까지만 입력 가능합니다")
-    float rating
+    BigDecimal rating
 ) {
 
 }
